@@ -1,6 +1,6 @@
 {
   "name": "@xaliksss/eslint-config",
-  "version": "2.0.1",
+  "version": "3.0.0",
   "main": "index.js",
   "keywords": [
     "eslint",
@@ -18,11 +18,12 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@eslint/js": "^9.0.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "globals": "^15.0.0"
+    "eslint-plugin-prettier": "^5.2.1",
+    "globals": "^15.9.0",
+    "typescript-eslint": "^8.2.0"
   },
   "peerDependencies": {
-    "eslint": "^9.0.0"
+    "eslint": "^9.9.0",
+    "prettier": "^3.3.3"
   }
 }
