@@ -1,1 +1,3 @@
-module.exports = require("./eslint.config.js");
+import config from "./eslint.config.js";
+
+export default config;
